Implemented Ratings page on student/TA view
Notes: will only show for the student if they are viewing the queue when they are removed. The form will only appear if the student is removed from the queue (not by leaving themselves). If a student rates the same session again, it will update the same record in the db, not add a new record.